fix inverted log message
When trying to follow a new primary, the type of the new primary is checked. When it is *not* of type primary, a warning is logged and the follow attempt is aborted. As far as I can tell, the message itself is exactly the opposite of what it should be: The node is not primary, thus it is standby / "in recovery". Feel free to correct me if I’m misunderstanding.