install-requirements.xml: fix compatibility matrix
The same document states that "repmgr 5.5.0 is compatible with all
supported PostgreSQL versions from 13.x", so 12 shouldn't be listed in
the matrix.

Moreover, 14 is missing in 5.5 and 5.4.1, which looks like an accidental
omission that might have happened when support for 15 was added on a
version that only advertized 13 up to that point. 14 should have been
added along with 15 then.